+ Made the grammarSystem a module * Added module grammarSystem to settings.gradle
* moved all files related to grammar into the module
