+++ mods/grammarSystem/src/main/java/org/terasology/grammarSystem/logic/grammar/shapes/complex/ComplexRule.java
@@ -1,0 +1,20 @@
+package org.terasology.grammarSystem.logic.grammar.shapes.complex;
+
+import org.terasology.grammarSystem.logic.grammar.shapes.Shape;
+
+import java.util.List;
+
+/** @author Tobias 'skaldarnar' Nett */
+public abstract class ComplexRule extends Shape {
+
+    /**
+     * Returns the successor elements of this complex shape.
+     * <p/>
+     * The successor elements of a complex shape are for example the arguments of a _divide_ or _repeat_ rule. The successor elements cannot
+     * be simply looked up in the org.terasology.logic.grammar rules but rather depend on the specific complex shape rule/command they
+     * belong to.
+     *
+     * @return the successor elements of the complex shape rule
+     */
+    abstract public List<Shape> getElements();
+}
