+++ mods/grammarSystem/src/main/java/org/terasology/grammarSystem/world/building/GrammarLoader.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2012 Benjamin Glatzel <benjamin.glatzel@me.com>
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.terasology.grammarSystem.world.building;
+
+import org.terasology.asset.AssetLoader;
+import org.terasology.asset.AssetUri;
+import org.terasology.grammarSystem.logic.grammar.assets.Grammar;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.net.URL;
+import java.util.List;
+
+/** @author skaldarnar */
+public class GrammarLoader implements AssetLoader<Grammar> {
+    @Override
+    public Grammar load(AssetUri uri, InputStream stream, List<URL> urls) throws IOException {
+        return null;  //To change body of implemented methods use File | Settings | File Templates.
+    }
+}
