+++ settings.gradle
@@ -1,2 +1,3 @@
 rootProject.name = 'Terasology'
-include 'mods:funnyBlocks', 'mods:fences', 'mods:core', 'mods:minerals', 'mods:miniion', 'mods:soils', 'mods:test', 'mods:books', 'mods:potions', 'mods:portals', 'mods:morelights', 'mods:dynamicBlocks', 'mods:hunger'
+include 'mods:funnyBlocks', 'mods:fences', 'mods:core', 'mods:minerals', 'mods:miniion', 'mods:soils', 'mods:test', 'mods:books',
+        'mods:potions', 'mods:portals', 'mods:morelights', 'mods:dynamicBlocks', 'mods:hunger', 'mods:grammarSystem'
