--- src/main/java/org/terasology/world/generator/building/Tree.java
@@ -1,21 +1,0 @@
-package org.terasology.world.generator.building;
-
-import org.terasology.logic.grammar.shapes.Shape;
-import org.terasology.model.structures.BlockCollection;
-
-import java.util.List;
-
-/**
- * @author Tobias 'Skaldarnar' Nett
- *         <p/>
- *         A basic composite pattern for derivation trees.
- */
-public abstract class Tree {
-    public abstract BlockCollection derive();
-
-    public abstract void setParent(Tree parent);
-
-    public abstract Shape getShape();
-
-    public abstract List<TreeNode> findActiveNodes();
-}
